devicemanager: 2.1.7, __init__: 2.0.2
Moved Slackbot back to devicemanager (couldn't import)
